requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
psycopg2
PyJWT